Calculator/calc/requirements.txt
Django==5.1.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.1
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
Jinja2==3.1.4
pillow==10.4.0
pip==24.3.1
selenium==4.23.1